Update README.rst
Hi, I added the parameter names in the example for clarification. Then I changed the order of the sections, to be more intuitive:
1. Create a public tenant with admin user
2. Create a Tenant and use the same admin user
3. Create subsequent users (can also be admin users if defined beforehand)

I changed the Paragraph about provisioning to a Note — it's considerate to mention, but the app will work if it's done in the request cycle.
Consequently, I promoted the Note about the User creation to a paragraph, since it is required and essential to the workings of the package.

Thanks for your great work. I hope this PR to your liking, else just close the request, no hard feelings.

Regards Max